Remove cache_item_ext from cache
